--- a/.github/workflows/bvt.yml
+++ b/.github/workflows/bvt.yml
@@ -3,7 +3,10 @@ on: [pull_request]
 jobs:
   make:
     name: Make
-    runs-on: ubuntu-latest
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest, macos-latest]
     steps:
       - name: Checkout
         uses: actions/checkout@v1
@@ -11,9 +14,13 @@ jobs:
         run: |
           make deps
           make
+
   clippy_check:
     name: Clippy Check
-    runs-on: ubuntu-latest
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest, macos-latest]
     steps:
       - uses: actions/checkout@v1
       - run: rustup component add clippy
@@ -21,4 +28,4 @@ jobs:
         with:
           token: ${{ secrets.GITHUB_TOKEN }}
           args: --all-features
-
+          name: checks-${{ matrix.os }}