--- a/Cargo.toml
+++ b/Cargo.toml
@@ -22,6 +22,8 @@ thiserror = "1.0"
 async-trait = { version = "0.1.31", optional = true }
 tokio = { version = "1", features = ["rt", "sync", "io-util", "macros", "time"], optional = true }
 futures = { version = "0.3", optional = true }
+
+[target.'cfg(target_os = "linux")'.dependencies]
 tokio-vsock = { version = "0.3.1", optional = true }
 
 [build-dependencies]