--- a/src/sync/server.rs
+++ b/src/sync/server.rs
@@ -27,6 +27,8 @@ use std::sync::{Arc, Mutex};
 use std::thread::JoinHandle;
 use std::{io, thread};
 
+#[cfg(not(target_os = "linux"))]
+use crate::common::set_fd_close_exec;
 use crate::common::{self, MESSAGE_TYPE_REQUEST};
 use crate::context;
 use crate::error::{get_status, Error, Result};
@@ -395,7 +397,7 @@ impl Server {
                         let pollers: &mut [libc::pollfd] = &mut pollers;
                         libc::poll(
                             pollers as *mut _ as *mut libc::pollfd,
-                            pollers.len() as u64,
+                            pollers.len() as _,
                             -1,
                         )
                     };
@@ -421,6 +423,7 @@ impl Server {
                         break;
                     }
 
+                    #[cfg(target_os = "linux")]
                     let fd = match accept4(listener, SockFlag::SOCK_CLOEXEC) {
                         Ok(fd) => fd,
                         Err(e) => {
@@ -428,6 +431,25 @@ impl Server {
                             break;
                         }
                     };
+
+                    // Non Linux platforms do not support accept4 with SOCK_CLOEXEC flag, so instead
+                    // use accept and call fcntl separately to set SOCK_CLOEXEC.
+                    // Because of this there is chance of the descriptor leak if fork + exec happens in between.
+                    #[cfg(not(target_os = "linux"))]
+                    let fd = match accept(listener) {
+                        Ok(fd) => {
+                            if let Err(err) = set_fd_close_exec(fd) {
+                                error!("fcntl failed after accept: {:?}", err);
+                                break;
+                            };
+                            fd
+                        }
+                        Err(e) => {
+                            error!("failed to accept error {:?}", e);
+                            break;
+                        }
+                    };
+
                     let methods = methods.clone();
                     let quit = Arc::new(AtomicBool::new(false));
                     let child_quit = quit.clone();