@mxpv 2021-11-06
I've added MacOS target to CI and it somehow confuses Github Actions. I found a couple suggestions [here](https://github.community/t/expected-waiting-for-status-to-be-reported/18001/2), but don't have access to repo settings.

![image](https://user-images.githubusercontent.com/865334/140618809-a295a772-57d3-4586-85f8-55cb4852b0cb.png)